Added source for the DSDT patcher. Included in the directory is my current version of the the dsd file for the dell. I have not modified the dsdt patcher to produce it yet.
Updated the ACPI module so that it follows the ACPI spec more closely, and works better with the dsdt file.This DOES NOT calculate a time (Not provided by bios) so it will remain at calculated in the battery manger. I'll fix that soon (or make it guess it like the old one did).

Some things I plan on doing is ad a new method like the _BST (in the dsdt file) to allow for extended information (such as cycle count, temperatures, etc). Once that is done, I may rename it to something else, (as it wasn't made by apple afaik, and it will do more than battery stuff).

Also, I have a bit of debugging info that needs to be removed from the dsdt file (not reporting design voltage, I was using it to read the other memory locations).

So what doesn't work:
Charging time remaining
Batt Serial number
Extended info (this isn't part of the acpi spec, however it shouldn't break anything if the old kext is use once I implement it)
DSDT patcher